refactor: move DocumentCollectionFactory to index module ICIJ/datashare#1283
