+++ datashare-app/src/main/java/org/icij/datashare/mode/CommonMode.java
@@ -36,8 +36,8 @@ import org.icij.datashare.extract.RedisUserDocumentQueue;
 import org.icij.datashare.extract.RedisUserReportMap;
 import org.icij.datashare.nlp.EmailPipeline;
 import org.icij.datashare.nlp.OptimaizeLanguageGuesser;
-import org.icij.datashare.tasks.DocumentCollectionFactory;
-import org.icij.datashare.tasks.MemoryDocumentCollectionFactory;
+import org.icij.datashare.extract.DocumentCollectionFactory;
+import org.icij.datashare.extract.MemoryDocumentCollectionFactory;
 import org.icij.datashare.tasks.TaskFactory;
 import org.icij.datashare.tasks.TaskManager;
 import org.icij.datashare.tasks.TaskManagerAmqp;
