+++ datashare-app/src/main/java/org/icij/datashare/tasks/DeduplicateTask.java
@@ -4,6 +4,7 @@ import com.google.inject.Inject;
 import com.google.inject.assistedinject.Assisted;
 import org.icij.datashare.PropertiesProvider;
 import org.icij.datashare.cli.DatashareCli;
+import org.icij.datashare.extract.DocumentCollectionFactory;
 import org.icij.datashare.user.User;
 import org.icij.extract.queue.DocumentQueue;
 import org.slf4j.Logger;
