+++ datashare-app/src/main/java/org/icij/datashare/web/StatusResource.java
@@ -16,7 +16,7 @@ import org.icij.datashare.PropertiesProvider;
 import org.icij.datashare.Repository;
 import org.icij.datashare.com.DataBus;
 import org.icij.datashare.openmetrics.StatusMapper;
-import org.icij.datashare.tasks.DocumentCollectionFactory;
+import org.icij.datashare.extract.DocumentCollectionFactory;
 import org.icij.datashare.text.indexing.Indexer;
 import org.icij.extract.queue.DocumentQueue;
 import org.slf4j.Logger;
