+++ datashare-app/src/test/java/org/icij/datashare/tasks/ExtractNlpTaskTest.java
@@ -1,12 +1,12 @@
 package org.icij.datashare.tasks;
 
 import org.icij.datashare.PropertiesProvider;
+import org.icij.datashare.extract.MemoryDocumentCollectionFactory;
 import org.icij.datashare.text.Document;
 import org.icij.datashare.text.Language;
 import org.icij.datashare.text.indexing.Indexer;
 import org.icij.datashare.text.nlp.AbstractPipeline;
 import org.icij.datashare.user.User;
-import org.icij.extract.queue.DocumentQueue;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
