+++ datashare-app/src/test/java/org/icij/datashare/tasks/IndexTaskTest.java
@@ -1,6 +1,7 @@
 package org.icij.datashare.tasks;
 
 import org.icij.datashare.PropertiesProvider;
+import org.icij.datashare.extract.DocumentCollectionFactory;
 import org.icij.datashare.text.indexing.elasticsearch.ElasticsearchSpewer;
 import org.icij.task.Options;
 import org.junit.Test;
